Moved "app" back to "sample"
