+++ sample/src/androidTest/java/com/raizlabs/universaladapter/ApplicationTest.java
