+++ sample/src/main/java/com/raizlabs/universaladapter/sample/AdapterActivity.java
@@ -1,4 +1,4 @@
-package com.raizlabs.universaladapter.app;
+package com.raizlabs.universaladapter.sample;
 
 import android.content.Context;
 import android.content.Intent;
