+++ sample/src/main/java/com/raizlabs/universaladapter/sample/MainActivity.java
@@ -1,4 +1,4 @@
-package com.raizlabs.universaladapter.app;
+package com.raizlabs.universaladapter.sample;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
