+++ sample/src/main/res/layout/activity_listview.xml
