+++ sample/src/main/res/layout/activity_recyclerview.xml
