+++ sample/src/main/res/layout/activity_viewpager.xml
