+++ sample/src/main/res/layout/list_item_adapter.xml
