+++ sample/src/main/res/layout/list_item_menu.xml
