+++ sample/src/main/res/layout/list_item_text.xml
